Спринт 14. Коммит после первого ревью
