+++ src/main/java/ru/practicum/shareit/item/controller/ItemController.java
@@ -1,30 +1,26 @@
 package ru.practicum.shareit.item.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.validateInterfaces.Create;
-import ru.practicum.shareit.validateInterfaces.Update;
 import ru.practicum.shareit.item.model.dto.CommentDto;
 import ru.practicum.shareit.item.model.dto.ItemDto;
 import ru.practicum.shareit.item.service.ItemService;
+import ru.practicum.shareit.validateInterfaces.Create;
+import ru.practicum.shareit.validateInterfaces.Update;
 
 import java.util.List;
 
 @RestController
 @RequestMapping("/items")
+@AllArgsConstructor
 public class ItemController {
     private final ItemService itemService;
 
-    @Autowired
-    public ItemController(ItemService itemService) {
-        this.itemService = itemService;
-    }
-
     @PostMapping("{itemId}/comment")
     @ResponseStatus(HttpStatus.OK)
-    public CommentDto createComment(@RequestBody @Validated CommentDto commentDto,
+    public CommentDto createComment(@RequestBody @Validated(Create.class) CommentDto commentDto,
                                     @RequestHeader("X-Sharer-User-Id") long userId,
                                     @PathVariable long itemId) {
         return itemService.createComment(commentDto, userId, itemId);
@@ -47,7 +43,8 @@ public class ItemController {
 
     @GetMapping("/{itemId}")
     @ResponseStatus(HttpStatus.OK)
-    public ItemDto getItem(@PathVariable long itemId, @RequestHeader("X-Sharer-User-Id") long userId) {
+    public ItemDto getItem(@PathVariable long itemId,
+                           @RequestHeader("X-Sharer-User-Id") long userId) {
         return itemService.getItem(itemId, userId);
     }
 
@@ -59,7 +56,8 @@ public class ItemController {
 
     @GetMapping("/search")
     @ResponseStatus(HttpStatus.OK)
-    public List<ItemDto> getItemByText(@RequestParam String text, @RequestHeader("X-Sharer-User-Id") long userId) {
+    public List<ItemDto> getItemByText(@RequestParam String text,
+                                       @RequestHeader("X-Sharer-User-Id") long userId) {
         return itemService.getItemsByText(text, userId);
     }
 }
