+++ src/main/java/ru/practicum/shareit/item/model/dto/CommentDto.java
@@ -3,8 +3,11 @@ package ru.practicum.shareit.item.model.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import ru.practicum.shareit.validateInterfaces.Create;
+import ru.practicum.shareit.validateInterfaces.Update;
 
 import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Size;
 import java.time.LocalDateTime;
 
 @Data
@@ -12,7 +15,8 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 public class CommentDto {
     private Long id;
-    @NotBlank
+    @NotBlank(groups = Create.class)
+    @Size(groups = {Create.class, Update.class}, max = 1000)
     private String text;
     private String authorName;
     private LocalDateTime created;
