+++ src/main/java/ru/practicum/shareit/item/model/dto/ItemDto.java
@@ -3,11 +3,13 @@ package ru.practicum.shareit.item.model.dto;
 import jdk.jfr.BooleanFlag;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import ru.practicum.shareit.validateInterfaces.Create;
 import ru.practicum.shareit.booking.model.dto.BookingDtoForItem;
+import ru.practicum.shareit.validateInterfaces.Create;
+import ru.practicum.shareit.validateInterfaces.Update;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
 import java.util.List;
 
 @Data
@@ -15,8 +17,10 @@ import java.util.List;
 public class ItemDto {
     private Long id;
     @NotBlank(groups = Create.class)
+    @Size(groups = {Create.class, Update.class}, max = 100)
     private String name;
     @NotBlank(groups = Create.class)
+    @Size(groups = {Create.class, Update.class}, max = 500)
     private String description;
     @BooleanFlag
     @NotNull(groups = Create.class)
