+++ src/main/java/ru/practicum/shareit/item/model/mapper/CommentMapperImpl.java
@@ -1,6 +1,6 @@
 package ru.practicum.shareit.item.model.mapper;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 import ru.practicum.shareit.booking.repository.BookingRepository;
 import ru.practicum.shareit.exception.BadRequestException;
@@ -9,8 +9,8 @@ import ru.practicum.shareit.item.model.Comment;
 import ru.practicum.shareit.item.model.Item;
 import ru.practicum.shareit.item.model.dto.CommentDto;
 import ru.practicum.shareit.item.repository.CommentRepository;
-import ru.practicum.shareit.user.repository.UserRepository;
 import ru.practicum.shareit.user.model.User;
+import ru.practicum.shareit.user.repository.UserRepository;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -18,18 +18,11 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Component
+@AllArgsConstructor
 public class CommentMapperImpl implements CommentMapper {
     private final BookingRepository bookingRepository;
     private final UserRepository userRepository;
     private final CommentRepository commentRepository;
-
-    @Autowired
-    public CommentMapperImpl(BookingRepository bookingRepository, UserRepository userRepository,
-                             CommentRepository commentRepository) {
-        this.bookingRepository = bookingRepository;
-        this.userRepository = userRepository;
-        this.commentRepository = commentRepository;
-    }
 
     @Override
     public Comment mapFrom(CommentDto entity, Item item, long userId) {
@@ -39,7 +32,8 @@ public class CommentMapperImpl implements CommentMapper {
                 .orElseThrow(() -> new BadRequestException("Вы не брали вещь в аренду"));
 
         User author = userRepository.findById(userId)
-                .orElseThrow(() -> new EntityNotFoundException("User not with id " + userId + " not found"));
+                .orElseThrow(() -> new EntityNotFoundException(String.format("Пользователь с идентификатором %d" +
+                        " не найден", userId)));
 
         return new Comment(entity.getId(), entity.getText(), item, author, currentTime);
     }
