+++ src/main/resources/schema.sql
@@ -11,8 +11,8 @@ CREATE TABLE users (
 CREATE TABLE items (
   id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
   user_id BIGINT NOT NULL,
-  name VARCHAR(255) NOT NULL,
-  description VARCHAR(1000) NOT NULL,
+  name VARCHAR(100) NOT NULL,
+  description VARCHAR(500) NOT NULL,
   available BOOLEAN NOT NULL,
   CONSTRAINT pk_item PRIMARY KEY (id),
   FOREIGN  KEY (user_id) REFERENCES users (id) ON DELETE CASCADE
@@ -32,7 +32,7 @@ CREATE TABLE bookings (
 
 CREATE TABLE comments (
     id BIGINT GENERATED BY DEFAULT AS IDENTITY NOT NULL,
-    text VARCHAR NOT NULL,
+    text VARCHAR(1000) NOT NULL,
     item_id BIGINT NOT NULL,
     author_id INTEGER NOT NULL,
     created TIMESTAMP WITHOUT TIME ZONE NOT NULL,
